backend/requirements.txt
# FastAPI and ASGI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
geoalchemy2>=0.14.0
psycopg2-binary>=2.9.0

# Validation and Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Authentication (JWT only, no password hashing needed for OTP login)
python-jose[cryptography]>=3.3.0

# HTTP Client (for SMS/Email APIs)
httpx>=0.26.0

# Form data
python-multipart>=0.0.6

# Email
aiosmtplib>=3.0.0
